backend/requirements.txt
# =============================================================================
# Network Copilot - Python Dependencies
# =============================================================================

# API Framework (replaces Chainlit)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0

# LLM Orchestration
langgraph>=1.0.0
langchain-core>=1.0.0

# LLM Providers (loaded lazily based on LLM_PROVIDER)
langchain-ollama>=1.0.0
langchain-openai>=1.0.0
langchain-anthropic>=1.0.0
langchain-google-genai>=4.0.0

# SSH / Network Automation
netmiko>=4.4.0
ntc-templates>=6.0.0

# Environment & Configuration
python-dotenv>=1.0.1
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
alembic>=1.14.1

# HTTP Client (used in tests and health checks)
requests>=2.32.3

# Testing
pytest>=9.0.0
